move ass drawing code to module
The `highlight_entry` function will probably be moved to a different
module at some point.
